ZeroDayArmor/CYBERSEC_AI_AGENT/requirements.txt
# CYBERSEC_AI_AGENT — LangChain + LangGraph + Streamlit
# Python: use 3.11 or 3.12 (create venv: python3.12 -m venv .venv). LangChain is not compatible with 3.14+ yet.
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-community>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.3.0
pydantic>=2.0,<3.0
python-dotenv>=1.0.0
requests>=2.28.0

# UI
streamlit>=1.28.0
pandas>=2.0.0
plotly
Pillow>=9.0.0

# RAG / embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
torchvision>=0.15.0

# Tests & Linting
pytest>=7.0.0
black
flake8

